se crea clase Producto y TesProducto Main || funciona ok

+++ Clases/src/com/krakedev/Auto.java
@@ -3,5 +3,5 @@ package com.krakedev;
 public class Auto {
 	String marca;
 	int anio;
-	double precio;
+	float precio;
 }
